requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
importlib_metadata==9.0.0
jsonpath-python==1.1.5
numpy==2.4.4
openrouter==0.9.1
opentelemetry-api==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
pandas==3.0.2
prometheus_client==0.25.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.1.0
PyYAML==6.0.3
six==1.17.0
sqlglot==30.6.0
structlog==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
zipp==3.23.1
